pyproject.toml
[project]
name = "papers"
version = "0.1.0"
description = "Centralized Knowledge Proxy with Semantic Cache"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "beaver-db>=1.0.0",
    "castor-io>=0.1.0",
    "httpx>=0.27.0",
    "fastembed>=0.2.0",
    "pydantic-settings>=2.13.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
